ft-#74: Refactor into hooks
