+++ src/pages/SendMoneyMobile/SendMoneyMobileData/index.tsx
@@ -2,12 +2,12 @@ import { useForm } from "react-hook-form";
 
 import Input from "@components/SendMoney/Input";
 import Tag from "@components/SendMoney/Tag";
+import { SEND_MONEY_MOBILE_BUNDLES } from "@constants/index";
 
 import NetworkSelector from "../NetworkSelector";
+import useCurrentBundleAmount from "../hooks/useCurrentBundleAmount";
 
 import type { TSendMoneyMobileNetwork } from "@customTypes/SendMoneyMobileNetwork";
-import { SEND_MONEY_MOBILE_BUNDLES } from "@constants/index";
-import { useEffect, useMemo, useState } from "react";
 
 type TSendMoneyMobileDataProps = {
   currentNetwork: TSendMoneyMobileNetwork;
@@ -17,37 +17,11 @@ type TSendMoneyMobileDataProps = {
   handleCurrentNetworkChange: (networkId: string) => void;
 };
 
-const useSendMoneyMobileData = () => {
-  const { watch, setValue, register, handleSubmit: _handleSubmit } = useForm();
-  const [currentBundleAmount, setCurrentBundleAmount] = useState<string>("");
-  const amountValue = watch("amount");
-
-  useEffect(() => {
-    if (!currentBundleAmount) {
-      setValue("amount", "");
-
-      return;
-    }
-
-    setValue("amount", currentBundleAmount);
-  }, [currentBundleAmount]);
-
-  const isPayButtonDisabled = useMemo(() => {
-    if (!amountValue) return true;
-
-    return false;
-  }, [amountValue]);
-
-  const handleBundleClick = (amount: string) => {
-    setCurrentBundleAmount(currentBundleAmount === amount ? "" : amount);
-  };
+const useSendMoneyMobileAirtime = () => {
+  const form = useForm();
 
   return {
-    currentBundleAmount,
-    isPayButtonDisabled,
-    handleBundleClick,
-    register,
-    handleSubmit: _handleSubmit,
+    form,
   };
 };
 
@@ -58,8 +32,9 @@ const SendMoneyMobileData = ({
   handleCurrentNetworkClick,
   handleCurrentNetworkChange,
 }: TSendMoneyMobileDataProps) => {
-  const { currentBundleAmount, isPayButtonDisabled, handleBundleClick, register, handleSubmit } =
-    useSendMoneyMobileData();
+  const { form } = useSendMoneyMobileAirtime();
+  const { currentBundleAmount, isPayButtonDisabled, handleBundleClick } = useCurrentBundleAmount(form);
+  const { register } = form;
 
   return (
     <div>
